Add WebDriverManagerWebDriverSupplier
The new WebDriverSupplier wraps the duplicate code from
DockerWebDriverSupplier and LocalWebDriverSupplier.

As it is useful in its own right, it is not abstract.
