+++ modules/selenium/src/main/java/org/shakespeareframework/selenium/DockerWebDriverSupplier.java
@@ -8,21 +8,15 @@ import org.openqa.selenium.WebDriver;
  * A {@link WebDriverSupplier} for Docker based browser instances. Uses {@link WebDriverManager} to manage the container
  * and create the {@link WebDriver}.
  */
-public class DockerWebDriverSupplier extends WebDriverSupplier {
-
-    private final WebDriverManager webDriverManager;
-    private WebDriver webDriver;
+public class DockerWebDriverSupplier extends WebDriverManagerWebDriverSupplier {
 
     /**
      * @param browserType            the {@link BrowserType} to be setup
      * @param additionalCapabilities additional {@link Capabilities} for the {@link WebDriver}
      */
     public DockerWebDriverSupplier(BrowserType browserType, Capabilities additionalCapabilities) {
-        super(browserType, additionalCapabilities);
-        this.webDriverManager = WebDriverManager
-                .getInstance(browserType.getWebDriverClass())
-                .capabilities(getCapabilities())
-                .browserInDocker();
+        super(WebDriverManager.getInstance(browserType.getWebDriverClass()).browserInDocker(),
+                browserType, additionalCapabilities);
     }
 
     /**
@@ -30,19 +24,5 @@ public class DockerWebDriverSupplier extends WebDriverSupplier {
      */
     public DockerWebDriverSupplier(BrowserType browserType) {
         this(browserType, null);
-    }
-
-    @Override
-    public WebDriver get() {
-        if (webDriver == null) {
-            webDriver = webDriverManager.create();
-        }
-        return webDriver;
-    }
-
-    @Override
-    public void close() {
-        webDriverManager.quit();
-        webDriver = null;
     }
 }
